requirements.txt
click
pydantic
python_dateutil